added to richRDD: writePartitions - richHConf: writePartition - switched over writeRows on RichRDDRegionValue - switched over writeBlocks on BlockMatrix
