+++ src/main/scala/is/hail/utils/richUtils/RichRDD.scala
@@ -1,9 +1,10 @@
 package is.hail.utils.richUtils
 
-import java.io.OutputStream
+import java.io.{Closeable, OutputStream}
 
 import is.hail.sparkextras.ReorderedPartitionsRDD
 import is.hail.utils._
+import org.apache.commons.lang3.StringUtils
 import org.apache.hadoop
 import org.apache.hadoop.io.compress.CompressionCodecFactory
 import org.apache.spark.{NarrowDependency, Partition, TaskContext}
@@ -31,7 +32,6 @@ class RichRDD[T](val r: RDD[T]) extends AnyVal {
 
     val codecFactory = new CompressionCodecFactory(hConf)
     val codec = Option(codecFactory.getCodec(new hadoop.fs.Path(filename)))
-    val headerExt = codec.map(_.getDefaultExtension).getOrElse("")
 
     hConf.delete(filename, recursive = true) // overwriting by default
 
@@ -64,7 +64,7 @@ class RichRDD[T](val r: RDD[T]) extends AnyVal {
       fatal("write failed: no success indicator found")
 
     if (!parallelWrite) {
-      hConf.copyMerge(parallelOutputPath, filename, rWithHeader.getNumPartitions, true, false)
+      hConf.copyMerge(parallelOutputPath, filename, rWithHeader.getNumPartitions, hasHeader = false)
     }
   }
 
@@ -167,7 +167,32 @@ class RichRDD[T](val r: RDD[T]) extends AnyVal {
       .subsetPartitions((0 to idxLast).toArray)
   }
   
-//  def writePartitions[S](makeStream: (OutputStream) => S, write: (S, Int, Iterator[T]) => ()) {
-//    
-//  }
+  def writePartitions[S <: Closeable](path: String, 
+    makeStream: (OutputStream) => S, write: (S, Int, Iterator[T]) => Long): Long = {
+    
+    val sc = r.sparkContext
+    val hadoopConf = sc.hadoopConfiguration
+    
+    hadoopConf.mkDir(path + "/parts")
+    
+    val sHadoopConfBc = sc.broadcast(new SerializableHadoopConfiguration(hadoopConf))
+    
+    val nPartitions = r.getNumPartitions
+    val d = digitsNeeded(nPartitions)
+
+    val itemCount = r.mapPartitionsWithIndex { case (i, it) =>
+      val is = i.toString
+      assert(is.length <= d)
+      val pis = StringUtils.leftPad(is, d, "0")
+
+      val filename = path + "/parts/part-" + pis
+
+      Iterator.single(sHadoopConfBc.value.value.writePartition(filename)(makeStream, i, it, write))
+    }
+      .fold(0L)(_ + _)
+
+    info(s"wrote $itemCount items in $nPartitions partitions")
+    
+    itemCount
+  }
 }
